requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.5.1
    # via
    #   -r requirements.in
    #   django
asttokens==2.0.5
    # via
    #   -r requirements.in
    #   stack-data
backcall==0.2.0
    # via
    #   -r requirements.in
    #   ipython
backports-zoneinfo==0.2.1
    # via
    #   -r requirements.in
    #   django
decorator==5.1.1
    # via
    #   -r requirements.in
    #   ipdb
    #   ipython
django==4.0.4
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
executing==0.8.3
    # via
    #   -r requirements.in
    #   stack-data
flake8==4.0.1
    # via -r requirements.in
ipdb==0.13.9
    # via -r requirements.in
ipython==8.3.0
    # via
    #   -r requirements.in
    #   ipdb
isort==5.10.1
    # via -r requirements.in
jedi==0.18.1
    # via
    #   -r requirements.in
    #   ipython
matplotlib-inline==0.1.3
    # via
    #   -r requirements.in
    #   ipython
mccabe==0.6.1
    # via
    #   -r requirements.in
    #   flake8
parso==0.8.3
    # via
    #   -r requirements.in
    #   jedi
pexpect==4.8.0
    # via
    #   -r requirements.in
    #   ipython
pickleshare==0.7.5
    # via
    #   -r requirements.in
    #   ipython
prompt-toolkit==3.0.29
    # via
    #   -r requirements.in
    #   ipython
ptyprocess==0.7.0
    # via
    #   -r requirements.in
    #   pexpect
pure-eval==0.2.2
    # via
    #   -r requirements.in
    #   stack-data
pycodestyle==2.8.0
    # via
    #   -r requirements.in
    #   flake8
pyflakes==2.4.0
    # via
    #   -r requirements.in
    #   flake8
pygments==2.12.0
    # via
    #   -r requirements.in
    #   ipython
pytz==2022.1
    # via
    #   -r requirements.in
    #   djangorestframework
six==1.16.0
    # via
    #   -r requirements.in
    #   asttokens
sqlparse==0.4.2
    # via
    #   -r requirements.in
    #   django
stack-data==0.2.0
    # via
    #   -r requirements.in
    #   ipython
toml==0.10.2
    # via
    #   -r requirements.in
    #   ipdb
traitlets==5.1.1
    # via
    #   -r requirements.in
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.5
    # via
    #   -r requirements.in
    #   prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
